# server/docs/swagger.yaml
definitions:
...
  model.AlipayPreCreatePayToFrontend:
    properties:
      qr_code:
        type: string
    type: object
...
  model.CasbinInfo:
    properties:
      casbinItems:
        items:
          $ref: '#/definitions/model.CasbinItem'
        type: array
      roleID:
        description: 权限id
        type: integer
    type: object
  model.CasbinItem:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
...
  model.EpayPreCreatePay:
    properties:
      money:
        description: "商品金额,如：1.00\t单位：元，最大2位小数"
        type: string
      name:
        description: 商品名称,如超过127个字节会自动截取
        type: string
      notify_url:
        description: 服务器异步通知地址
        type: string
      out_trade_no:
        description: 商户订单号
        type: string
      pid:
        description: 商户ID
        type: integer
      return_url:
        description: 页面跳转通知地址
        type: string
      sign:
        description: 签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign
          = md5 ( a=b&c=d&e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。
        type: string
      sign_type:
        description: 签名类型
        type: string
      type:
        description: 支付方式
        type: string
    type: object
  model.EpayPreCreatePayToFrontend:
    properties:
      epay_api_url:
        type: string
      epay_pre_create_pay:
        $ref: '#/definitions/model.EpayPreCreatePay'
    type: object
  model.EpayResultResponse:
    properties:
      money:
        description: 商品金额
        type: string
      name:
        description: 商品名称
        type: string
      out_trade_no:
        description: 商户订单号
        type: string
      param:
        description: 业务扩展参数
        type: string
      pid:
        description: 商户ID
        type: integer
      sign:
        description: 签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign
          = md5 ( a=b&c=d&e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。
        type: string
      sign_type:
        description: 签名类型
        type: string
      trade_no:
        description: 易支付订单号
        type: string
      trade_status:
        description: 支付状态
        type: string
      type:
        description: 支付方式
        type: string
    type: object
  model.FieldParamsItem:
    properties:
      condition:
        description: = > < <> like
        type: string
      condition_value:
        type: string
      field:
        type: string
      field_type:
        type: string
      operator:
        description: AND OR
        type: string
    type: object
...
  model.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      component:
        type: string
      created_at:
        type: string
      id:
        type: integer
      meta:
        $ref: '#/definitions/model.Meta'
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      remarks:
        type: string
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      updated_at:
        type: string
    type: object
  model.Meta:
    properties:
      icon:
        type: string
      isAffix:
        type: boolean
      isHide:
        type: boolean
      isIframe:
        type: boolean
      isKeepAlive:
        type: boolean
      isLink:
        type: string
      title:
        type: string
    type: object
...
  model.Order:
    properties:
      balance_amount:
        type: string
      buyer_logon_id:
        type: string
      buyer_pay_amount:
        type: string
      coupon_amount:
        type: string
      coupon_id:
        type: integer
      coupon_name:
        type: string
      created_at:
        type: string
      customer_service_id:
        description: 服务参数
        type: integer
      deliver_text:
        type: string
      deliver_type:
        type: string
      des:
        type: string
      duration:
        type: integer
      goods_id:
        description: 商品参数
        type: integer
      goods_type:
        type: string
      id:
        type: integer
      order_remarks:
        type: string
      order_type:
        type: string
      original_amount:
        type: string
      out_trade_no:
        type: string
      pay_id:
        description: 支付参数
        type: integer
      pay_info:
        allOf:
        - $ref: '#/definitions/model.PreCreatePayToFrontend'
        description: 支付信息，epay，alipay等"
      pay_type:
        type: string
      price:
        type: string
      subject:
        type: string
      total_amount:
        type: string
      trade_no:
        type: string
      trade_status:
        type: string
      updated_at:
        type: string
      user_id:
        description: 关联用户
        type: integer
      user_name:
        type: string
    type: object
  model.Pagination:
    properties:
      order_by:
        type: string
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
...
  model.PreCreatePayToFrontend:
    properties:
      alipay_info:
        $ref: '#/definitions/model.AlipayPreCreatePayToFrontend'
      epay_info:
        $ref: '#/definitions/model.EpayPreCreatePayToFrontend'
    type: object
...
  model.QueryParams:
    properties:
      field_params_list:
        items:
          $ref: '#/definitions/model.FieldParamsItem'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/model.Pagination'
        description: 分页参数
      table_name:
        type: string
    required:
    - table_name
    type: object
...
  model.Role:
    properties:
      casbins:
        items:
          $ref: '#/definitions/model.CasbinItem'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      role_name:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
      user_group:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      balance:
        type: number
      created_at:
        type: string
      enable:
        type: boolean
      enable_email:
        type: boolean
      enable_tg_bot:
        description: 通知参数
        type: boolean
      enable_web_mail:
        type: boolean
      id:
        type: integer
      invitation_code:
        type: string
      nick_name:
        type: string
      orders:
        description: 订单，has many
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
      referrer_user_id:
        type: integer
      role_group:
        description: 关联参数
        items:
          $ref: '#/definitions/model.Role'
        type: array
      tg_id:
        type: integer
      updated_at:
        type: string
      user_name:
        type: string
      when_balance_changed:
        type: boolean
      when_purchased:
        type: boolean
      when_service_almost_expired:
        type: boolean
    type: object
...
  response.ResponseStruct:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact:
    url: https://github.com/ppoonk/AirGo
  description: AirGo前后分离,多用户代理面板
  license:
    name: GPL v3.0
    url: https://github.com/ppoonk/AirGo/blob/main/LICENSE
  title: AirGo
paths:
...
  /api/admin/role/delRole:
    delete:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 删除角色
      tags:
      - '[admin api] role'
  /api/admin/role/getAllPolicy:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取全部权限
      tags:
      - '[admin api] role'
  /api/admin/role/getPolicyByID:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CasbinInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取权限列表ByRoleId
      tags:
      - '[admin api] role'
  /api/admin/role/getRoleList:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取角色列表
      tags:
      - '[admin api] role'
  /api/admin/role/newRole:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 新建角色
      tags:
      - '[admin api] role'
  /api/admin/role/updateRole:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 修改角色信息
      tags:
      - '[admin api] role'
...
  /api/customer/finance/getBalanceStatementList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取余额流水
      tags:
      - '[customer api] finance'
  /api/customer/finance/getCommissionStatementList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取佣金流水
      tags:
      - '[customer api] finance'
  /api/customer/finance/getCommissionSummary:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取佣金统计
      tags:
      - '[customer api] finance'
  /api/customer/finance/getInvitationUserList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取邀请人数
      tags:
      - '[customer api] finance'
  /api/customer/finance/withdrawToBalance:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 提现
      tags:
      - '[customer api] finance'
...
  /api/public/shop/epayNotify:
    get:
      parameters:
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.EpayResultResponse'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            type: string
        "400":
          description: 请求错误
      summary: 易支付异步回调
      tags:
      - '[public api] shop'
...
swagger: "2.0"